+++ homeassistant/components/sharkiq/strings.json
@@ -1,13 +1,6 @@
 {
-  "title": "Shark IQ",
   "config": {
     "step": {
-      "init": {
-        "data": {
-          "username": "[%key:common::config_flow::data::username%]",
-          "password": "[%key:common::config_flow::data::password%]"
-        }
-      },
       "user": {
         "data": {
           "username": "[%key:common::config_flow::data::username%]",
